moved user tests to separate classes
